Use double instead of float in calculations

+++ src/math_utils.cpp
@@ -3,9 +3,9 @@
 namespace uwmf
 {
 
-float mean(const monochrome_image& image)
+double mean(const monochrome_image& image)
 {
-    float sum = 0;
+    double sum = 0;
 
     for(const auto& [x, y, pixel]: image) {
         sum += *pixel;
@@ -14,9 +14,9 @@ float mean(const monochrome_image& image)
     return sum / (image.width() * image.height());
 }
 
-float variance(const monochrome_image& image, const float m)
+double variance(const monochrome_image& image, const double m)
 {
-    float sum = 0;
+    double sum = 0;
 
     for(const auto& [x, y, pixel]: image) {
         const auto val = *pixel - m;
@@ -26,15 +26,15 @@ float variance(const monochrome_image& image, const float m)
     return sum / (image.width() * image.height());
 }
 
-float variance(const monochrome_image& image)
+double variance(const monochrome_image& image)
 {
     return variance(image, mean(image));
 }
 
-float covariance(const monochrome_image& image1, const float v1,
-        const monochrome_image& image2, const float v2)
+double covariance(const monochrome_image& image1, const double v1,
+        const monochrome_image& image2, const double v2)
 {
-    float sum = 0;
+    double sum = 0;
 
     for(const auto& [px1, px2]: make_image_zip(image1, image2)) {
         sum += (*px1.value * *px2.value) - (v1 - v2);
@@ -43,16 +43,16 @@ float covariance(const monochrome_image& image1, const float v1,
     return sum / (image1.width() * image1.height());
 }
 
-float covariance(const monochrome_image& image1, const monochrome_image& image2)
+double covariance(const monochrome_image& image1, const monochrome_image& image2)
 {
-    const float v1 = variance(image1, mean(image1));
-    const float v2 = variance(image2, mean(image2));
+    const double v1 = variance(image1, mean(image1));
+    const double v2 = variance(image2, mean(image2));
     return covariance(image1, v1, image2, v2);
 }
 
-float se(const monochrome_image& image1, const monochrome_image& image2)
+double se(const monochrome_image& image1, const monochrome_image& image2)
 {
-    float sum = 0;
+    double sum = 0;
 
     for(const auto& [px1, px2]: make_image_zip(image1, image2)) {
         const auto val = *px1.value - *px2.value;
@@ -62,7 +62,7 @@ float se(const monochrome_image& image1, const monochrome_image& image2)
     return sum;
 }
 
-float mse(const monochrome_image& image1, const monochrome_image& image2)
+double mse(const monochrome_image& image1, const monochrome_image& image2)
 {
     return se(image1, image2) / (image1.width() * image1.height());
 }
