+++ src/metrics.h
@@ -7,9 +7,9 @@
 namespace uwmf
 {
 
-float psnr(const monochrome_image& original, const monochrome_image& restored);
-float ief(const monochrome_image& original, const monochrome_image& restored,
+double psnr(const monochrome_image& original, const monochrome_image& restored);
+double ief(const monochrome_image& original, const monochrome_image& restored,
         const monochrome_image& noisy);
-float ssim(const monochrome_image& original, const monochrome_image& restored);
+double ssim(const monochrome_image& original, const monochrome_image& restored);
 
 } // uwmf
